requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-Limiter==4.1.1
Flask-Talisman==1.1.0
Flask-Caching==2.3.1
marshmallow==4.2.2
python-dotenv==1.2.2
email-validator==2.3.0
pytest==9.0.2
pytest-cov==7.0.0
black==26.3.0
flake8==7.3.0
mypy==1.19.1
